machine-learning-client/requirements.txt
opencv-python==4.8.0.76
numpy==1.24.3
pymongo==4.5.0
tensorflow==2.13.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
pylint==2.17.4 